Set chat as home view and stabilize theming
