+++ src/app/sign-in/[[...sign-in]]/page.tsx
@@ -2,8 +2,8 @@ import { SignIn } from "@clerk/nextjs";
 
 export default function Page() {
   return (
-    <div className="flex justify-center items-center h-screen">
-      <SignIn />
+    <div className="flex h-screen items-center justify-center">
+      <SignIn redirectUrl="/" afterSignInUrl="/" />
     </div>
   );
 }
