+++ src/app/sign-up/[[...sign-up]]/page.tsx
@@ -2,8 +2,8 @@ import { SignUp } from "@clerk/nextjs";
 
 export default function Page() {
   return (
-    <div className="flex justify-center items-center h-screen">
-      <SignUp />
+    <div className="flex h-screen items-center justify-center">
+      <SignUp redirectUrl="/" afterSignUpUrl="/" />
     </div>
   );
 }
