+++ src/components/Sidebar.tsx
@@ -31,7 +31,10 @@ export default function Sidebar({
   currentThreadId: string | null;
   className?: string;
 }) {
-  const { theme, setTheme } = useTheme();
+  const { resolvedTheme, setTheme } = useTheme();
+  const [mounted, setMounted] = useState(false);
+
+  useEffect(() => setMounted(true), []);
 
   // ✅ Replace any[] with Thread[]
   const [threads, setThreads] = useState<Thread[]>([]);
@@ -124,15 +127,15 @@ export default function Sidebar({
 
   return (
     <aside
-      className={`flex h-full w-[280px] shrink-0 flex-col border-r border-white/10 bg-[#050507] text-white ${className}`.trim()}
+      className={`flex h-full w-[280px] shrink-0 flex-col border-r border-slate-200/70 bg-white text-slate-900 transition-colors duration-300 dark:border-white/10 dark:bg-[#050507] dark:text-white ${className}`.trim()}
     >
-      <div className="flex items-center gap-2 px-5 py-4 border-b border-white/10">
-        <div className="flex h-10 w-10 items-center justify-center rounded-full bg-emerald-500/20 text-emerald-300">
+      <div className="flex items-center gap-2 border-b border-slate-200/70 bg-slate-50 px-5 py-4 transition-colors duration-300 dark:border-white/10 dark:bg-white/5">
+        <div className="flex h-10 w-10 items-center justify-center rounded-full bg-emerald-500/15 text-emerald-600 dark:text-emerald-300">
           <Sparkles size={20} />
         </div>
         <div className="flex-1">
-          <p className="text-xs uppercase tracking-wide text-white/60">Workspace</p>
-          <h2 className="font-semibold text-white">SPGPT</h2>
+          <p className="text-xs uppercase tracking-wide text-slate-500 dark:text-white/60">Workspace</p>
+          <h2 className="font-semibold text-slate-900 dark:text-white">SPGPT</h2>
         </div>
         <UserButton />
       </div>
@@ -140,7 +143,7 @@ export default function Sidebar({
       <div className="px-4 py-3">
         <button
           onClick={createThread}
-          className="flex w-full items-center justify-center gap-2 rounded-xl bg-white/10 px-3 py-2 text-sm font-medium text-white transition hover:bg-white/20"
+          className="flex w-full items-center justify-center gap-2 rounded-xl bg-slate-900/5 px-3 py-2 text-sm font-medium text-slate-900 transition hover:bg-slate-900/10 dark:bg-white/10 dark:text-white dark:hover:bg-white/20"
         >
           <Plus size={16} />
           New chat
@@ -149,7 +152,7 @@ export default function Sidebar({
 
       <div className="flex-1 space-y-1 overflow-y-auto px-3 pb-6">
         {threads.length === 0 ? (
-          <div className="rounded-xl border border-dashed border-white/10 bg-white/5 p-4 text-center text-xs text-white/60">
+          <div className="rounded-xl border border-dashed border-slate-200/70 bg-white p-4 text-center text-xs text-slate-500 transition-colors duration-300 dark:border-white/10 dark:bg-white/5 dark:text-white/60">
             Start your first conversation to see it here.
           </div>
         ) : (
@@ -161,8 +164,8 @@ export default function Sidebar({
                 key={thread._id}
                 className={`group flex items-center gap-2 rounded-xl px-3 py-2 text-sm transition ${
                   isActive
-                    ? "bg-emerald-500/15 text-emerald-200"
-                    : "text-white/70 hover:bg-white/10 hover:text-white"
+                    ? "bg-emerald-500/15 text-emerald-700 dark:text-emerald-200"
+                    : "text-slate-600 hover:bg-slate-900/5 hover:text-slate-900 dark:text-white/70 dark:hover:bg-white/10 dark:hover:text-white"
                 }`}
               >
                 <button
@@ -185,18 +188,18 @@ export default function Sidebar({
 
                 <div className="flex items-center gap-1 opacity-0 transition group-hover:opacity-100">
                   {editingId === thread._id ? (
-                    <button onClick={() => saveRename(thread._id)} className="rounded p-1 hover:bg-white/10">
+                    <button onClick={() => saveRename(thread._id)} className="rounded p-1 hover:bg-slate-900/5 dark:hover:bg-white/10">
                       <Check size={15} />
                     </button>
                   ) : (
                     <button
                       onClick={() => startRename(thread._id, thread.title)}
-                      className="rounded p-1 hover:bg-white/10"
+                      className="rounded p-1 hover:bg-slate-900/5 dark:hover:bg-white/10"
                     >
                       <Edit2 size={15} />
                     </button>
                   )}
-                  <button onClick={() => deleteThread(thread._id)} className="rounded p-1 hover:bg-white/10">
+                  <button onClick={() => deleteThread(thread._id)} className="rounded p-1 hover:bg-slate-900/5 dark:hover:bg-white/10">
                     <Trash2 size={15} />
                   </button>
                 </div>
@@ -206,16 +209,23 @@ export default function Sidebar({
         )}
       </div>
 
-      <div className="mt-auto flex items-center justify-between px-4 py-4 border-t border-white/10">
-        <div className="text-xs text-white/50">
-          <p className="font-medium text-white/70">Theme</p>
+      <div className="mt-auto flex items-center justify-between border-t border-slate-200/70 px-4 py-4 transition-colors duration-300 dark:border-white/10">
+        <div className="text-xs text-slate-500 dark:text-white/50">
+          <p className="font-medium text-slate-700 dark:text-white/70">Theme</p>
           <p>Switch between light and dark</p>
         </div>
         <button
-          onClick={() => setTheme(theme === "light" ? "dark" : "light")}
-          className="rounded-xl border border-white/10 bg-white/5 p-2 transition hover:bg-white/10"
+          onClick={() =>
+            setTheme((resolvedTheme ?? "system") === "dark" ? "light" : "dark")
+          }
+          className="rounded-xl border border-slate-200/70 bg-slate-900/5 p-2 transition hover:bg-slate-900/10 dark:border-white/10 dark:bg-white/5 dark:hover:bg-white/10"
+          aria-label="Toggle theme"
         >
-          {theme === "light" ? <Moon size={18} /> : <Sun size={18} />}
+          {mounted && (resolvedTheme ?? "light") === "dark" ? (
+            <Sun size={18} />
+          ) : (
+            <Moon size={18} />
+          )}
         </button>
       </div>
     </aside>
