+++ src/providers/ThemeProvider.tsx
@@ -5,11 +5,10 @@ import { ReactNode, useEffect, useState } from "react";
 export default function Providers({ children }: { children: ReactNode }) {
   const [mounted, setMounted] = useState(false);
   useEffect(() => setMounted(true), []);
-  if (!mounted) return <>{children}</>;   // avoid hydration mismatch
 
   return (
-    <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
-      {children}
+    <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
+      {mounted ? children : <div className="opacity-0">{children}</div>}
     </ThemeProvider>
   );
 }
